Refactor

+++ backend/src/test/java/io/github/projektalmanac/amoxtli/backend/service/UserServiceTest.java
@@ -239,7 +239,7 @@ class UserServiceTest {
 
     @Test
     void verificaCorreo() {
-        long userId = 1;
+        int userId = 1;
 
         //////// Codigo valido.///////
         CodigoVerificacionDto codigoVerificacionDto = new CodigoVerificacionDto();
